app/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pgvector
python-dotenv
openai
sentence-transformers
alembic
psycopg2-binary
redis
pydantic
transformers
torch

